decimal a float

+++ src/AppForSEII2526.API/DTOs/BocadilloDTO.cs
@@ -7,7 +7,7 @@
         public string TipoPanNombre { get; set; }
 
         [Precision(10, 2)]
-        public decimal PVP { get; set; }
+        public float PVP { get; set; }
         public int Id { get; set; }
 
         public BocadilloDTO()
@@ -15,7 +15,7 @@
 
         }
 
-        public BocadilloDTO(int id, string nombre, string tipoPanNombre, Tamaño tamaño, decimal PVP)
+        public BocadilloDTO(int id, string nombre, string tipoPanNombre, Tamaño tamaño, float PVP)
         {
             Id = id;
             Nombre = nombre;
